+++ src/Components/Game/Game.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import './Game.css';
 import Board from '/Users/knethil/projects/my-first-extension/src/Components/Board/Board';
+import {calculateWinner,getWinningSquares} from '/Users/knethil/projects/my-first-extension/src/Utils/helper';
 
 class Game extends React.Component {
   constructor(props) {
@@ -143,44 +144,3 @@ class Game extends React.Component {
 }
 
 export default Game;
-
-// ========================================
-
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6]
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
-}
-
-function getWinningSquares(squares){
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6]
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return lines[i];
-    }
-  }
-}
